Move camera to track top of tower
Technically we're actually moving everything _else_ down, but the
effect is the same. Once we lock in a piece that ends up above the
halfway point of the screen height we shift everything down.

We also can then stop updating and drawing any minos that are off the
bottom of the screen. This is a big deal for performance as our
framerate started to drop after processing ~250 minos and went down to
30FPS after ~500 minos.

We can now play with thousands of minos (I tested it), as long as no
more than a couple hundered are on screen at once. This is fine
because there's not enough space to have too many on screen even if
you fill the whole width of the game. The camera moves them down an it
remains manageable.

Maybe someone with a vertical monitor could run into problems if they
try hard enough :P

We've also added backgrounds for the main menu and the level scene,
these move when the minos move but only a few pixels at a time giving
us a nice parallax effect as we ascend our tower.

Also added a debug mode (toggle by pressing `d`) which hows the
current framerate and how many minos are in hot or cold storage.