Add outro scene
First we've added a new `outro` scene which copies over most of the
current sprites from the `level-01` scene. It also create new mino
sprites based off the info in `:off-screen-minos`. It positions
everything and then starts a couple of animations using tweens.

First we fade all the minos to black and slowly start fading the
background to white (this is done by having a transparent white
rectangle sprite covering the background slowly tween up in opacity).

Then after a delay we start to tween all of the minos so the zoom
upwards giving us the effect of panning the camera down the tower to
where we started. At the same time we tween the background
image (still visible under the semi-transparent background-cover) so
that the parallax effect from the game happens in reverse.

It's not particularly clean, but it works and will do for now.